Added pt1 and pt2.
Pt1 is about introduction, basic plotting, scatterplot, line plot, bar charts and dejunkifying a graph. This will help u to get used to this library and follow with the other parts; 
Pt2 is about subplots, histograms, boxplots, heatmaps, animation and interactivity. There is a lot of more advenced stuff in here, but it is just a quickly introduction to other types of graphs that you will probably need to interpretate if your need is to do some statistics with data.